api/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
PyYAML==6.0.1
GitPython==3.1.41
openai==1.12.0
python-dotenv==1.0.0
pydantic==2.6.0
sentence-transformers==2.3.0
chromadb==0.4.22
